Fix stylelint and eslint errors
